disable save() and save-as() functions due to deprecation
legacy support is assumed to be dropped in 4.1.

Due to recent updates, several save() and save-as() routines have been
removed from the codebase, however, several editors were still calling the
routines.

Because I'm not confident enough to simply delete the lines, they are
simply commented out for now, but based on my understanding, it sounds
like the eventual plan is for these routines to either be:

1. repurposed (using the serialization methods)
2. deleted completely